feat: annotations + tours feature flags with lazy-loaded feature chunks
Adds a features block to instance.json (absent keys default to off) read
via src/config/features.js. Everything annotation-related in Map.svelte
moves into a new MapAnnotations.svelte, which inserts its OL layer at
the same stacking position and reports list visibility through
mapState.annotationsListShowing so the MapControls gating is unchanged.

MapAnnotations, TourController, and TourListModal are now fetched via
dynamic import only when their flag is on (and, for tours, on first
use), which also drops supabase-js and the annotation code out of the
initial bundle: entry chunk 1,445 kB -> 1,135 kB (422 -> 335 kB gzip).

Entry points gated: Research-tab annotation buttons, splash tour button,
tour-list modal, and view:tour share URLs (which fall back to the splash
when tours are off). The Research tab hides entirely when annotations
are off and the instance defines no bbox search functions.

Verified in dev preview with flags on (annotate, load annotations, tour
list, tour playback, tour share URL) and off (entry points absent, zero
feature modules fetched, tour URLs degrade gracefully).
