+++ src/lib/state.svelte.js
@@ -34,6 +34,7 @@ export const mapState = $state({
     annotationRead: false,
     annotationEntry: false,
     annotationSave: false,
+    annotationsListShowing: false,
     layerChangePopup: false,
     center: null,
     zoom: null,
